feat: PR 14.6.9 — Trend Persistence & Governance Forecasting
- run_snapshot() persists 3 FaFreshnessTrendSnapshot rows (7d/30d/90d) per
  run with score_delta, fresh_delta, expired_delta, coverage_risk_delta vs
  period baseline; each write wrapped in try/except so failure never blocks
  the daily snapshot commit
- get_trend_history(): paginated retrieval of stored trend snapshots by period
- get_forecast(): linear velocity projection (velocity = delta/days) to
  30/60/90d horizons, clamped [0,100]; early_warning flag when score will
  breach threshold within horizon_days
- 2 new routes: GET /freshness/trends/history, GET /freshness/forecast
- Fix P2: get_evidence_history() now applies since_date to paginated query so
  days parameter is respected; fallback returns empty window (not out-of-window
  rows) when evidence exists but has no data in the requested window
- Fix P2: _delta_for_period() adds lower_bound (days*2) constraint so a
  60d-old snapshot cannot serve as a 7d baseline; same fix applied to
  coverage_risk_delta in get_cgin_trends()
- Fix pre-existing: compute_score_delta() now returns float() to satisfy
  -> float annotation when called with integer arguments
- No new tables or migrations; fa_freshness_trend_snapshots existed in 0132
- Governance: route-inventory-generate, refresh_contract_authority,
  generate_platform_inventory, soc-review-sync all pass; fg-fast 316s

+++ services/freshness_score_history/models.py
@@ -34,4 +34,4 @@ def compute_trend_direction(score_delta: float) -> TrendDirection:
 
 
 def compute_score_delta(current: float, baseline: float) -> float:
-    return round(current - baseline, 2)
+    return float(round(current - baseline, 2))
